+++ plugins/frontend/backstage-plugin-datadog/src/alpha/apis.ts
@@ -1,19 +1,14 @@
-import {
-  ApiBlueprint,
-  createApiFactory,
-  discoveryApiRef,
-} from '@backstage/frontend-plugin-api';
+import { ApiBlueprint, discoveryApiRef } from '@backstage/frontend-plugin-api';
 import { datadogApiRef, DatadogApiClient } from '../api';
 
 /**
  * @alpha
  */
 export const datadogApi = ApiBlueprint.make({
-  params: {
-    factory: createApiFactory({
+  params: paramsFactory =>
+    paramsFactory({
       api: datadogApiRef,
       deps: { discoveryApi: discoveryApiRef },
       factory: ({ discoveryApi }) => new DatadogApiClient({ discoveryApi }),
     }),
-  },
 });