+++ plugins/frontend/backstage-plugin-github-insights/src/components/Widgets/LanguagesCard/LanguagesCard.tsx
@@ -34,7 +34,7 @@ import { GitHubAuthorizationWrapper } from '@roadiehq/github-auth-utils-react';
 
 const useStyles = makeStyles(theme => ({
   infoCard: {
-    marginBottom: theme.spacing(3),
+    marginBottom: '10px',
     '& + .MuiAlert-root': {
       marginTop: theme.spacing(3),
     },